:recycle: moving directory for HdDatepicker component
